Add own printf
Extracting decimal digits:
This was interesting, no 32-bit divions are allowed!
We can neither do something like 32-bit * 32-bit,
nor 32-bit / 32-bit, or the compiler will generate
calls to unresolved symbols __U4D, __U4M, ...
The trick is to use a division table, and instead
of dividing, we subtract as long as the dividend is
larger or equal the divisor.

Signed-off-by: Andreas J. Reichel <andreas@reichel.bayern>